Update mCAP_main.m
In the event of a binary seed mask loaded into the inputs, this prevents  the seed_vol to be binary and let's accept double values.

+++ mCAP_main.m
@@ -44,7 +44,7 @@ while isconverged==0 && it <= max_iteration
     seed_map=CAP_norm{it}(seedmask,:);
 
     
-    newseed=repmat(seedmask,1,ik); %% the new seed has ik parcels
+    newseed=repmat(double(seedmask),1,ik); %% the new seed has ik parcels
     newseed(seedmask==1,:)=seed_map;
